feat: show all schol scans in table
