+++ apps/web/src/routes/scans/[scan_id]/+page.ts
